Similarities between associated genes in drug compounds
After compiling information on the genes associated with specific drug
compounds, a jaccard was calculated for each pair of drugs. Additionally, the
drugs were categorized in four subsets: target, carrier, transporter and enzyme.
The combined jaccard values for each category were compared on a univariate
level using a histogram and on a bivariate level with a hexbin scatterplot. The
compound similarity data for each pair of drugs was added as a fifth subset
to compare the chemical makeup of the drugs in relation to the categories.
A multi-lined pointplot was also created to demonstrate the relationship between
compound similarity and mean jaccard value. Using 0.5 as a baseline value, the
compound similarity values were converted into contingency values and used with
the category jaccard values to create contingency tables for each pair of drug
compounds. Finally, the contingency relative frequency for each combination of
category was illustrated with five barplots.